+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- module: initial version of module `bitcoin-zeromq-client-bitcoin-kmp`
 
 ### Breaking
 - upgrade: update spring-boot from v2.7.6 to v3.0.2
@@ -22,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - upgrade: update docker lightninglabs/lnd from v0.15.3-beta to v0.15.5-beta
 - upgrade: update docker getumbrel/btc-rpc-explorer from v2.1.0 to v3.3.0
 
+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
 
 ## [0.5.0] - 2022-11-29
 ### Breaking